+++ packages/terra-clinical-detail-view/tests/jest/DetailView.test.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import DetailView from '../../src/DetailView';
 
+jest.mock('uuid', () => ({ v4: () => '00000000-0000-0000-0000-000000000000' }));
+
 const defaultVariety = <DetailView />;
 
 // Snapshot Tests
@@ -23,12 +25,12 @@ it('should render a title', () => {
 
 it('should set the default heading level of the title to h2', () => {
   const wrapper = shallow(<DetailView title="Title" />);
-  expect(wrapper.html('.primary-text')).toContain('<h2 class="primary-text">Title</h2>');
+  expect(wrapper.html('.primary-text')).toContain('<h2 class="primary-text" id="title-id-00000000-0000-0000-0000-000000000000">Title</h2>');
 });
 
 it('should set the heading level for the title', () => {
   const wrapper = shallow(<DetailView level={3} title="Title" />);
-  expect(wrapper.html('.primary-text')).toContain('<h3 class="primary-text">Title</h3>');
+  expect(wrapper.html('.primary-text')).toContain('<h3 class="primary-text" id="title-id-00000000-0000-0000-0000-000000000000">Title</h3>');
 });
 
 it('should render a smaller title', () => {