+++ packages/terra-clinical-result/src/terra-dev-site/doc/example/clinical-result-blood-pressure/ClinicalResultBloodPressureExtraDisplays.jsx
@@ -9,7 +9,7 @@ const bloodpressureResultValue = {
       value: '120',
       unit: 'mmHg',
     },
-    conceptDisplay: 'Blood Pressure Systolic',
+    conceptDisplay: 'Systolic',
     datetimeDisplay: 'Nov 23, 2019 13:31:31',
   },
   diastolic: {
@@ -18,9 +18,10 @@ const bloodpressureResultValue = {
       value: '80',
       unit: 'mmHg',
     },
-    conceptDisplay: 'Blood Pressure Diastolic',
-    datetimeDisplay: 'Nov 23, 2019 13:31:44',
+    conceptDisplay: 'Diastolic',
+    datetimeDisplay: 'Nov 23, 2019 13:31:31',
   },
+  isBloodPressureGrouped: true,
 };
 
 export default () => <ClinicalResultBloodPressure {...bloodpressureResultValue} />;