+++ packages/terra-clinical-result/tests/jest/_BloodPressureDisplay.test.jsx
@@ -1,4 +1,6 @@
 import React from 'react';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { shallowWithIntl } from 'terra-enzyme-intl';
 import BloodPressureDisplay from '../../src/_BloodPressureDisplay';
 import { DefaultBloodPressureResult } from '../../src/terra-dev-site/test/clinical-result/TestResults';
 
@@ -15,12 +17,12 @@ describe('BloodPressureDisplay', () => {
   const type = 'Systolic';
 
   it('should render a default BloodPressureDisplay', () => {
-    const display = shallow(<BloodPressureDisplay result={defaultResult} id={id} type={type} />);
+    const display = shallowWithIntl(<BloodPressureDisplay result={defaultResult} id={id} type={type} />).dive();
     expect(display).toMatchSnapshot();
   });
 
   it('should render a ResultError', () => {
-    const display = shallow(<BloodPressureDisplay id={id} type={type} />);
+    const display = shallowWithIntl(<BloodPressureDisplay id={id} type={type} />).dive();
     expect(display).toMatchSnapshot();
   });
 
@@ -29,7 +31,7 @@ describe('BloodPressureDisplay', () => {
       ...defaultResult,
       noData: true,
     };
-    const display = shallow(<BloodPressureDisplay result={result} id={id} type={type} />);
+    const display = shallowWithIntl(<BloodPressureDisplay result={result} id={id} type={type} />).dive();
     expect(display).toMatchSnapshot();
   });
 
@@ -39,17 +41,17 @@ describe('BloodPressureDisplay', () => {
       statusInError: true,
       interpretation: 'critical',
     };
-    const display = shallow(<BloodPressureDisplay result={result} id={id} type={type} />);
+    const display = shallowWithIntl(<BloodPressureDisplay result={result} id={id} type={type} />).dive();
     expect(display).toMatchSnapshot();
   });
 
   it('should hide the unit if hideUnit is true', () => {
-    const display = shallow(<BloodPressureDisplay result={defaultResult} id={id} type={type} hideUnit />);
+    const display = shallowWithIntl(<BloodPressureDisplay result={defaultResult} id={id} type={type} hideUnit />).dive();
     expect(display).toMatchSnapshot();
   });
 
   it('should hide the unit if data\'s cleanedUnit is equal to diastolicUnit prop', () => {
-    const display = shallow(<BloodPressureDisplay result={defaultResult} id={id} type={type} diastolicUnit={defaultResult.cleanedUnit} />);
+    const display = shallowWithIntl(<BloodPressureDisplay result={defaultResult} id={id} type={type} diastolicUnit={defaultResult.cleanedUnit} />).dive();
     expect(display).toMatchSnapshot();
   });
 
@@ -59,7 +61,7 @@ describe('BloodPressureDisplay', () => {
       statusInError: true,
       interpretation: 'critical',
     };
-    const display = shallow(<BloodPressureDisplay result={result} id={id} type={type} diastolicUnit={defaultResult.cleanedUnit} />);
+    const display = shallowWithIntl(<BloodPressureDisplay result={result} id={id} type={type} diastolicUnit={defaultResult.cleanedUnit} />).dive();
     expect(display).toMatchSnapshot();
   });
 });